requirements.txt
streamlit==1.28.0
psycopg2-binary==2.9.9
pandas==2.3.3
plotly==5.18.0

